+++ sys/dev/netif/my/if_my.c
@@ -872,7 +872,7 @@ my_attach(device_t dev)
 	sc->my_ldata_ptr = kmalloc(sizeof(struct my_list_data) + 8,
 				  M_DEVBUF, M_WAITOK);
 	sc->my_ldata = (struct my_list_data *) sc->my_ldata_ptr;
-	round = (unsigned int)sc->my_ldata_ptr & 0xF;
+	round = (uintptr_t)sc->my_ldata_ptr & 0xF;
 	roundptr = sc->my_ldata_ptr;
 	for (i = 0; i < 8; i++) {
 		if (round % 8) {
